requirements.txt
streamlit
pandas
plotly
plotly-calplot
Pillow
